requirements.txt
pandas>=2.0.0,<3.0.0
pyyaml>=6.0,<7.0
databricks-sql-connector>=3.0.0,<4.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
prefect==3.3.4
ruff>=0.1.0